+++ internal/resources/metal/project_ssh_key/models.go
@@ -1,4 +1,4 @@
-package metal_project_ssh_key
+package project_ssh_key
 
 import (
 	"path"
